Add --hostnameonly option to `floaty list --active`
Adds a new option to `floaty list`: `--hostnameonly` which causes
`floaty list --active` to output active VM hostnames, one per line,
with no additional information. This is useful for tooling that may
need to consume the output, particularly now that the default output
for the ABS backend has fields in a different order than the other
backends.

This commit also refactors the unit tests for the
`Utils.pretty_print_hosts` method and adds tests for the ABS backend, as
well as for the new options.

Lastly, this fixes a bug in the `Utils.pretty_print_hosts` method where
the `print_to_stderr` argument had no effect.